refactor tests into separate files for each suite
and make CallbackInfra tests narrower/ specific purpose per test.
